requirements.txt
pytorch~=1.10.1
tensorboard~=2.7.0
numpy~=1.21.2
attrdict~=2.0.1
